+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - feat: Refactor `GfFormField` field settings, choices, and inputs to use GraphQL interfaces.
 - feat: Deprecate `FormsConnectionOrderbyInput.field` in favor of `FormsConnectionOrderbyInput.column`.
 - fix: ensure latest mutation input data is used to prepare the field values on update mutations.
+- fix!: Change GraphQL field FormQuizConfirmation.isAutoformatted from type String to type Boolean.
+- fix!: Change GraphQL field FormQuizConfirmation.message from type Int to type String.
+- fix: Fix resolver for GfForm.quiz returning empty data.
 - dev!: Move `TypeRegistry` classes to `WPGraphQL\GF\Registry` namespace.
 - dev!: Register each GraphQL type on its own `add_action()` call.
 - dev!: Remove nullable `$type_registry` param from `Registrable::register()` interface method.
@@ -19,6 +22,7 @@
 - dev: Deprecate the `graphql_gf_form_field_setting_properties` filter in favor of `graphql_gf_form_field_setting_fields`.
 - dev: Deprecate the `graphql_gf_form_field_value_properties` filter in favor of `graphql_gf_form_field_value_fields`.
 - chore: Refactor `FormsConnectionResolver` to use new `AbstractConnectionResolver` methods.
+- test: Add basic WPUnit tests for GFForm.quiz fields.
 
 ## v0.11.4